refactor: group state primitives
