+++ src/state/profilesManager/profilesManager.ts
@@ -1,6 +1,5 @@
 import { createEvent, createStore } from 'effector';
 import { ProfileObject } from '@core/storage/ProfilesManager';
-import { NewProfile } from '@features/WorkspaceManager/ProfileCreator';
 
 export const createProfilesManagerApi = () => {
 	const $profiles = createStore<null | ProfileObject[]>(null);
@@ -14,9 +13,4 @@ export const createProfilesManagerApi = () => {
 		$profiles,
 		events,
 	};
-};
-
-export type ProfilesManagerApi = {
-	profiles: ProfileObject[] | null;
-	createProfile: (profile: NewProfile) => Promise<void>;
 };
